Use namespacing for consistency
The larger PHP world has embraced namespacing, and so should we.

Paves the way for autoloading, as per PSR-4.

This includes capitalization changes to class files, as classes should
always be capitalized, including their filenames
