+++ src/Alcatraz.php
@@ -1,17 +1,18 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-Apt-Get
+Alcatraz
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
 	
-	private $id = 'apt-get';
+	private $id = 'alcatraz';
 	private $kind = 'packages'; // for none found msg
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
@@ -25,8 +26,8 @@ class Repo {
 		$this->cache = new Cache();
 		
 		// get DB here if not dynamic search
-		//$data = (array) $this->cache->get_db($this->id);
-		//$this->pkgs = $data;
+		$data = (array) $this->cache->get_db($this->id)->packages;
+		$this->pkgs = $data;
 	}
 	
 	// return id | url | pkgstr
@@ -36,17 +37,16 @@ class Repo {
 	
 	function check($pkg, $query) {
 		if (!$query) { return true; }
-		if (strpos($pkg["name"], $query) !== false) {
+		if (strpos($pkg->name, $query) !== false) {
 			return true;
-		} else if (strpos($pkg["description"], $query) !== false) {
+		} else if (strpos($pkg->description, $query) !== false) {
 			return true;
-		} 
-	
+		}
 		return false;
 	}
 	
 	function search($query) {
-		if ( strlen($query) < $this->min_query_length) {
+		if ( strlen($query) < $this->min_query_length ) {
 			if ( strlen($query) === 0 ) { return; }
 			$this->cache->w->result(
 				"{$this->id}-min",
@@ -58,33 +58,31 @@ class Repo {
 			return;
 		}
 		
-		$this->pkgs = $this->cache->get_query_regex($this->id, $query, 'https://apps.ubuntu.com/cat/search/?q='.$query, '/<tr>([\s\S]*?)<\/tr>/i');
-		
-		foreach($this->pkgs as $item) {
-			preg_match('/<p>(.*?)<\/p>/i', $item, $matches);
-			$name = trim(strip_tags($matches[1]));
-			
-			preg_match('/<h3>([\s\S]*?)<\/h3>/i', $item, $matches);
-			$description = trim(strip_tags($matches[1]));
-		
-			$this->cache->w->result(
-				$name,
-				$this->makeArg($name, 'https://apps.ubuntu.com/cat/applications/'.$name, "*"),
-				$name,
-				$description,
-				"icon-cache/{$this->id}.png"
-			);
-			//break;
-			// only search till max return reached
-			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
-				break;
+		foreach($this->pkgs as $pkg ) {
+			// plugins, color_scheme, project_templates, file_templates
+			for( $i = 0; $i < count($pkg); $i++ ) {
+				
+				if ($this->check($pkg[$i], $query)) {
+					$this->cache->w->result(
+						$pkg[$i]->url,
+						$this->makeArg($pkg[$i]->name, $pkg[$i]->url, "*"),
+						$pkg[$i]->name,
+						$pkg[$i]->description,
+						"icon-cache/{$this->id}.png"
+					);
+				}
+				
+				// only search till max return reached
+				if ( count ( $this->cache->w->results() ) == $this->max_return ) {
+					break;
+				}
 			}
 		}
 		
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"https://apps.ubuntu.com/cat/search/?q={$query}",
+				"http://alcatraz.io//{$query}", // UPDATE NEEDED
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
@@ -93,12 +91,15 @@ class Repo {
 	}
 	
 	function xml() {
+		
+		
+		
 		$this->cache->w->result(
-			"{$this->id}-www",
-			"https://apps.ubuntu.com/cat/",
-			"Go to the website",
-			"https://apps.ubuntu.com",
-			"icon-cache/{$this->id}.png"
+			$this->id.'-www',
+			'http://alcatraz.io//',
+			'Go to the website',
+			'http://alcatraz.io/',
+			"icon-cache/".$this->id.".png"
 		);
 		
 		return $this->cache->w->toxml();
@@ -108,9 +109,11 @@ class Repo {
 
 // ****************
 
-/*$query = "leaflet";
+/*
+$query = "a";
 $repo = new Repo();
 $repo->search($query);
-echo $repo->xml();*/
+echo $repo->xml();
+*/
 
 ?>
