+++ src/Brew.php
@@ -1,4 +1,5 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
 Brew
@@ -7,7 +8,7 @@ Brew
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
