+++ src/Cocoa.php
@@ -1,18 +1,19 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-Ruby Gems
+Cocoa
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
 	
-	private $id = 'gems';
-	private $kind = 'gems'; // for none found msg
+	private $id = 'cocoa';
+	private $kind = 'libraries'; // for none found msg
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
 	
@@ -25,8 +26,8 @@ class Repo {
 		$this->cache = new Cache();
 		
 		// get DB here if not dynamic search
-		//$data = (array) $this->cache->get_db($this->id);
-		//$this->pkgs = $data;
+		$data = (array) $this->cache->get_db($this->id);
+		$this->pkgs = $data;
 	}
 	
 	// return id | url | pkgstr
@@ -35,10 +36,10 @@ class Repo {
 	}
 	
 	function check($pkg, $query) {
-		if (   !$query
-			|| strpos($pkg->name, $query) !== false
-			|| strpos($pkg->info, $query) !== false
-		) {
+		if (!$query) { return true; }
+		if (strpos($pkg->name, $query) !== false) {
+			return true;
+		} else if (isset($pkg->summary) && strpos($pkg->summary, $query) !== false) {
 			return true;
 		} 
 	
@@ -58,24 +59,31 @@ class Repo {
 			return;
 		}
 		
-		$this->pkgs = $this->cache->get_query_json(
-			$this->id, 
-			$query, 
-			"https://rubygems.org/api/v1/search?query={$query}"
-		);
+		//$this->pkgs = $this->cache->get_query_json($this->id, $query, "http://cocoadocs.org/?q=you/{$query}");
 		
 		foreach($this->pkgs as $pkg) {
-			if ($this->check($pkg, $query)) {
+			
+			// make params
+			if ($this->check($pkg,  $query)) {
 				$title = $pkg->name;
+				if (isset($pkg->main_version)) { $title .= ' ('.$pkg->main_version.')'; }
+				if (isset($pkg->user)) { $title .= ' ~ '.$pkg->user; }
+				
+				$url = (isset($pkg->url)) ? $$pkg->url : $pkg->doc_url;
+				$details = (isset($pkg->summary)) ? $pkg->summary : $pkg->framework;
+				
+				$icon = (isset($pkg->url)) ? "xcode" : "{$this->id}";
 				
 				$this->cache->w->result(
+					$pkg->name,
+					$this->makeArg($pkg->name, $url, "*"),
 					$title,
-					$this->makeArg($title, $pkg->project_uri, '*'),
-					$title,
-					$pkg->info,
-					"icon-cache/{$this->id}.png"
+					$details,
+					"icon-cache/{$icon}.png"
 				);
 			}
+			
+			
 			
 			// only search till max return reached
 			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
@@ -86,7 +94,7 @@ class Repo {
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"https://rubygems.org/search?utf8=%E2%9C%93&query={$query}",
+				"http://cocoadocs.org/?q={$query}",
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
@@ -95,11 +103,14 @@ class Repo {
 	}
 	
 	function xml() {
+		
+		
+		
 		$this->cache->w->result(
 			"{$this->id}-www",
-			'https://rubygems.org/',
+			'_TEMPLATE_URL_/',
 			'Go to the website',
-			'https://rubygems.org',
+			'_TEMPLATE_URL_',
 			"icon-cache/{$this->id}.png"
 		);
 		
@@ -110,11 +121,9 @@ class Repo {
 
 // ****************
 
-/*
-$query = "leaflet";
+/*$query = "leaflet";
 $repo = new Repo();
 $repo->search($query);
-echo $repo->xml();
-*/
+echo $repo->xml();*/
 
 ?>
