+++ src/Docker.php
@@ -1,49 +1,51 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-Hex
+docker
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
-
-	private $id = 'hex';
-	private $kind = 'components'; // for none found msg
+	
+	private $id = 'docker';
+	private $kind = 'images';
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
-
+	
 	private $cache;
 	private $w;
 	private $pkgs;
-
+	
 	function __construct() {
-
+		
 		$this->cache = new Cache();
-
+		
 		// get DB here if not dynamic search
-		//$data = (array) $this->cache->get_db('.$this->id.');
+		//$data = (array) $this->cache->get_db($this->id);
+		//$this->pkgs = $data;
 	}
-
+	
 	// return id | url | pkgstr
 	function makeArg($id, $url, $version) {
-		return $id . "|" . $url . "|" . "\"$id\":\"$version\",";
+		return $id . "|" . $url . "|" . $id;//"\"$id\":\"$version\",";
 	}
-
+	
 	/*function check($pkg, $query) {
 		if (!$query) { return true; }
 		if (strpos($pkg["name"], $query) !== false) {
 			return true;
 		} else if (strpos($pkg["description"], $query) !== false) {
 			return true;
-		}
-
+		} 
+	
 		return false;
 	}*/
-
+	
 	function search($query) {
 		if ( strlen($query) < $this->min_query_length) {
 			if ( strlen($query) === 0 ) { return; }
@@ -56,44 +58,57 @@ class Repo {
 			);
 			return;
 		}
-
-		$this->pkgs = $this->cache->get_query_json('hex', $query, 'https://hex.pm/api/packages?search='.$query);
-
+		
+		$this->pkgs = $this->cache->get_query_regex($this->id, $query, 'https://registry.hub.docker.com/search?q='.$query, '/(<a href="\/u\/.*"><div class="repo-list-item box">[\s\S]*?<\/a>)/i', 1);
+		
 		foreach($this->pkgs as $pkg) {
-			$url = str_replace("api/", "", $pkg->url);
+			
+			// make params
+			preg_match('/<a href="(.*?)">[\s\S]*?<h2>([\s\S]*?)<[\s\S]*<\/h2>([\s\S]*?)<\/div>/i', $pkg, $matches);
+			$title = trim(preg_replace('/\s+/', '', $matches[2]));
+			$url = 'https://registry.hub.docker.com'. trim(preg_replace('/\s+/', '', $matches[1]));
+			$description = trim(preg_replace('/\s+/', ' ', $matches[3]));
+			if (!$description || trim(preg_replace('/\s+/', '', $matches[3])) == '') {
+				$description = $url;
+			}
+		
+			preg_match('/<span class="timesince" data-time=".*">\s?([\s\S]*?)\s?<\/span>/i', $pkg, $matches);
+			$updated = trim(preg_replace('/\s+/', ' ', $matches[1]));
+	
 			$this->cache->w->result(
-				$pkg->url,
-				$this->makeArg($pkg->name, $url, "*"),
-				$pkg->name, $url,
+				$title,
+				$this->makeArg($title, $url, "*"),
+				$title.' ~ '.$updated,
+				$description,
 				"icon-cache/{$this->id}.png"
 			);
-
+			
 			// only search till max return reached
 			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
 				break;
 			}
 		}
-
+		
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"https://hex.pm/packages?search={$query}",
+				"https://registry.hub.docker.com/search?q={$query}",
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
 			);
 		}
 	}
-
+	
 	function xml() {
 		$this->cache->w->result(
 			"{$this->id}-www",
-			"http://hex.pm/",
-			"Go to the website",
-			"http://hex.pm",
+			'http://www.docker.io/',
+			'Go to the website',
+			'http://www.docker.io',
 			"icon-cache/{$this->id}.png"
 		);
-
+		
 		return $this->cache->w->toxml();
 	}
 
@@ -101,9 +116,11 @@ class Repo {
 
 // ****************
 
-/*$query = "leaflet";
+/*
+$query = "ng";
 $repo = new Repo();
 $repo->search($query);
-echo $repo->xml();*/
+echo $repo->xml();
+*/
 
 ?>
