+++ src/Gems.php
@@ -1,18 +1,19 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-grunt
+Ruby Gems
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
 	
-	private $id = 'grunt';
-	private $kind = 'plugins'; // for none found msg
+	private $id = 'gems';
+	private $kind = 'gems'; // for none found msg
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
 	
@@ -25,8 +26,8 @@ class Repo {
 		$this->cache = new Cache();
 		
 		// get DB here if not dynamic search
-		$data = (array) $this->cache->get_db($this->id)->aaData;
-		$this->pkgs = $data;
+		//$data = (array) $this->cache->get_db($this->id);
+		//$this->pkgs = $data;
 	}
 	
 	// return id | url | pkgstr
@@ -37,7 +38,7 @@ class Repo {
 	function check($pkg, $query) {
 		if (   !$query
 			|| strpos($pkg->name, $query) !== false
-			|| strpos($pkg->ds, $query) !== false
+			|| strpos($pkg->info, $query) !== false
 		) {
 			return true;
 		} 
@@ -58,30 +59,24 @@ class Repo {
 			return;
 		}
 		
-		//$this->pkgs = $this->cache->get_query_json($this->id, $query, "_TEMPLATE_SEARCH_URL_{$query}");
+		$this->pkgs = $this->cache->get_query_json(
+			$this->id, 
+			$query, 
+			"https://rubygems.org/api/v1/search?query={$query}"
+		);
 		
 		foreach($this->pkgs as $pkg) {
-			
-			// make params
 			if ($this->check($pkg, $query)) {
-				$title = str_replace('grunt-', '', $pkg->name); // remove grunt- from title
-			
-				// add author to title
-				if (isset($pkg->author)) {
-					$title .= " by " . $pkg->author;
-				}
-				$url = 'https://www.npmjs.org/package/' . $pkg->name;
+				$title = $pkg->name;
 				
-				//if (strpos($plugin->description, "DEPRECATED") !== false) { continue; } // skip DEPRECATED repos
 				$this->cache->w->result(
-					$pkg->name,
-					$this->makeArg($pkg->name, $url, "*"),
 					$title,
-					$pkg->ds,
+					$this->makeArg($title, $pkg->project_uri, '*'),
+					$title,
+					$pkg->info,
 					"icon-cache/{$this->id}.png"
 				);
 			}
-			
 			
 			// only search till max return reached
 			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
@@ -92,7 +87,7 @@ class Repo {
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"http://gruntjs.com/plugins/{$query}",
+				"https://rubygems.org/search?utf8=%E2%9C%93&query={$query}",
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
@@ -103,9 +98,9 @@ class Repo {
 	function xml() {
 		$this->cache->w->result(
 			"{$this->id}-www",
-			'http://gruntjs.com/',
+			'https://rubygems.org/',
 			'Go to the website',
-			'http://gruntjs.com',
+			'https://rubygems.org',
 			"icon-cache/{$this->id}.png"
 		);
 		
@@ -117,7 +112,7 @@ class Repo {
 // ****************
 
 /*
-$query = 'contrib';
+$query = "leaflet";
 $repo = new Repo();
 $repo->search($query);
 echo $repo->xml();
