+++ src/Gulp.php
@@ -1,18 +1,19 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-docker
+gulp
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
 	
-	private $id = 'docker';
-	private $kind = 'images';
+	private $id = 'gulp';
+	private $kind = 'plugins'; // for none found msg
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
 	
@@ -25,8 +26,8 @@ class Repo {
 		$this->cache = new Cache();
 		
 		// get DB here if not dynamic search
-		//$data = (array) $this->cache->get_db($this->id);
-		//$this->pkgs = $data;
+		$data = (array) $this->cache->get_db($this->id)->results;
+		$this->pkgs = $data;
 	}
 	
 	// return id | url | pkgstr
@@ -34,16 +35,16 @@ class Repo {
 		return $id . "|" . $url . "|" . $id;//"\"$id\":\"$version\",";
 	}
 	
-	/*function check($pkg, $query) {
+	function check($pkg, $query) {
 		if (!$query) { return true; }
-		if (strpos($pkg["name"], $query) !== false) {
+		if (strpos($pkg->name, $query) !== false) {
 			return true;
-		} else if (strpos($pkg["description"], $query) !== false) {
+		} else if (strpos($pkg->description, $query) !== false) {
 			return true;
 		} 
 	
 		return false;
-	}*/
+	}
 	
 	function search($query) {
 		if ( strlen($query) < $this->min_query_length) {
@@ -58,29 +59,34 @@ class Repo {
 			return;
 		}
 		
-		$this->pkgs = $this->cache->get_query_regex($this->id, $query, 'https://registry.hub.docker.com/search?q='.$query, '/(<a href="\/u\/.*"><div class="repo-list-item box">[\s\S]*?<\/a>)/i', 1);
+		//$this->pkgs = $this->cache->get_query_json($this->id, $query, "_TEMPLATE_SEARCH_URL_{$query}");
 		
 		foreach($this->pkgs as $pkg) {
 			
 			// make params
-			preg_match('/<a href="(.*?)">[\s\S]*?<h2>([\s\S]*?)<[\s\S]*<\/h2>([\s\S]*?)<\/div>/i', $pkg, $matches);
-			$title = trim(preg_replace('/\s+/', '', $matches[2]));
-			$url = 'https://registry.hub.docker.com'. trim(preg_replace('/\s+/', '', $matches[1]));
-			$description = trim(preg_replace('/\s+/', ' ', $matches[3]));
-			if (!$description || trim(preg_replace('/\s+/', '', $matches[3])) == '') {
-				$description = $url;
+			if ($this->check($pkg, $query)) {
+				$title = str_replace('gulp-', '', $pkg->name); // remove pulp- from title
+			
+				// add version to title
+				if (isset($pkg->version)) {
+					$title .= ' v'.$pkg->version;
+				}
+				// add author to title
+				if (isset($pkg->author)) {
+					$title .= " by " . $pkg->author;
+				}
+				
+				//if (strpos($plugin->description, "DEPRECATED") !== false) { continue; } // skip DEPRECATED repos
+				$this->cache->w->result(
+					$pkg->name,
+					$this->makeArg($pkg->name, $pkg->homepage, "*"),
+					$title,
+					$pkg->description,
+					"icon-cache/{$this->id}.png"
+				);
 			}
-		
-			preg_match('/<span class="timesince" data-time=".*">\s?([\s\S]*?)\s?<\/span>/i', $pkg, $matches);
-			$updated = trim(preg_replace('/\s+/', ' ', $matches[1]));
-	
-			$this->cache->w->result(
-				$title,
-				$this->makeArg($title, $url, "*"),
-				$title.' ~ '.$updated,
-				$description,
-				"icon-cache/{$this->id}.png"
-			);
+			
+			
 			
 			// only search till max return reached
 			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
@@ -91,7 +97,7 @@ class Repo {
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"https://registry.hub.docker.com/search?q={$query}",
+				"http://gulpjs.com/plugins/#?q={$query}",
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
@@ -102,9 +108,9 @@ class Repo {
 	function xml() {
 		$this->cache->w->result(
 			"{$this->id}-www",
-			'http://www.docker.io/',
+			'http://gulpjs.com/',
 			'Go to the website',
-			'http://www.docker.io',
+			'http://gulpjs.com',
 			"icon-cache/{$this->id}.png"
 		);
 		
@@ -116,7 +122,7 @@ class Repo {
 // ****************
 
 /*
-$query = "ng";
+$query = "min";
 $repo = new Repo();
 $repo->search($query);
 echo $repo->xml();
