+++ src/Maven.php
@@ -1,18 +1,19 @@
 <?php
+namespace WillFarrell\AlfredPkgMan;
 
 /*
-yo
+maven
 
 */
 
 // ****************
 
-require_once('cache.php');
+require_once('Cache.php');
 
 class Repo {
 	
-	private $id = 'yo';
-	private $kind = 'generators'; // for none found msg
+	private $id = 'maven';
+	private $kind = 'libraries'; // for none found msg
 	private $min_query_length = 1; // increase for slow DBs
 	private $max_return = 25;
 	
@@ -25,8 +26,8 @@ class Repo {
 		$this->cache = new Cache();
 		
 		// get DB here if not dynamic search
-		$data = (array) $this->cache->get_db($this->id);
-		$this->pkgs = $data;
+		//$data = (array) $this->cache->get_db($this->id);
+		//$this->pkgs = $data;
 	}
 	
 	// return id | url | pkgstr
@@ -34,16 +35,16 @@ class Repo {
 		return $id . "|" . $url . "|" . $id;//"\"$id\":\"$version\",";
 	}
 	
-	function check($pkg, $query) {
+	/*function check($pkg, $query) {
 		if (!$query) { return true; }
-		if (strpos($pkg->name, $query) !== false) {
+		if (strpos($pkg["name"], $query) !== false) {
 			return true;
-		} else if (strpos($pkg->description, $query) !== false) {
+		} else if (strpos($pkg["description"], $query) !== false) {
 			return true;
-		}
+		} 
 	
 		return false;
-	}
+	}*/
 	
 	function search($query) {
 		if ( strlen($query) < $this->min_query_length) {
@@ -58,28 +59,22 @@ class Repo {
 			return;
 		}
 		
-		//$this->pkgs = $this->cache->get_query_json($this->id, $query, "_TEMPLATE_SEARCH_URL_{$query}");
+		$this->pkgs = (array) $this->cache->get_query_json($this->id, $query, 'http://search.maven.org/solrsearch/select?q='.$query.'&rows=10&wt=json')->response->docs;
 		
 		foreach($this->pkgs as $pkg) {
 			
 			// make params
-			if ($this->check($pkg, $query)) {
-				$title = $pkg->name;
-				
-				// add author to title
-				if (isset($pkg->owner)) {
-					$title .= " by " . $pkg->owner;
-				}
-				
-				$this->cache->w->result(
-					$pkg->name,
-					$this->makeArg($pkg->name, $pkg->website, "*"),
-					$title,
-					$pkg->description,
-					"icon-cache/{$this->id}.png"
-				);
-			}
-			
+			$title = $pkg->a.' ('.$pkg->latestVersion.')';
+			$url = 'http://search.maven.org/#artifactdetails%7C'.$pkg->g.'%7C'.$pkg->a.'%7C'.$pkg->latestVersion.'%7C'.$pkg->p;
+			$details = 'GroupId: '.$pkg->id;
+	
+			$this->cache->w->result(
+				$pkg->a,
+				$this->makeArg($pkg->a, $url, "*"),
+				$title,
+				$details,
+				"icon-cache/{$this->id}.png"
+			);
 			
 			// only search till max return reached
 			if ( count ( $this->cache->w->results() ) == $this->max_return ) {
@@ -90,7 +85,7 @@ class Repo {
 		if ( count( $this->cache->w->results() ) == 0) {
 			$this->cache->w->result(
 				"{$this->id}-search",
-				"http://yeoman.io/community-generators.html?q={$query}",
+				"http://mvnrepository.com/search.html?query={$query}",
 				"No {$this->kind} were found that matched \"{$query}\"",
 				"Click to see the results for yourself",
 				"icon-cache/{$this->id}.png"
@@ -101,9 +96,9 @@ class Repo {
 	function xml() {
 		$this->cache->w->result(
 			"{$this->id}-www",
-			'http://yeoman.io/',
+			'http://mvnrepository.com/',
 			'Go to the website',
-			'http://yeoman.io',
+			'http://mvnrepository.com',
 			"icon-cache/{$this->id}.png"
 		);
 		
@@ -115,7 +110,7 @@ class Repo {
 // ****************
 
 /*
-$query = "ang";
+$query = "j";
 $repo = new Repo();
 $repo->search($query);
 echo $repo->xml();
